# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lalezar Online Shop API",
        "description": "Interactive documentation for the API endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/admin/cities": {
            "post": {
                "tags": [
                    "Admin - Cities"
                ],
                "summary": "Create a city",
                "operationId": "adminCitiesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "state_id",
                                    "name"
                                ],
                                "properties": {
                                    "state_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Karaj"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_capital": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CityResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/cities/{city}": {
            "patch": {
                "tags": [
                    "Admin - Cities"
                ],
                "summary": "Update a city",
                "operationId": "adminCitiesUpdate",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "state_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_capital": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CityResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/comments/{comment}/release": {
            "post": {
                "tags": [
                    "Admin - Comments"
                ],
                "summary": "Publish a comment",
                "operationId": "adminCommentsRelease",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment published",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/comments/{comment}/answer": {
            "post": {
                "tags": [
                    "Admin - Comments"
                ],
                "summary": "Answer a comment",
                "operationId": "adminCommentsAnswer",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answer"
                                ],
                                "properties": {
                                    "answer": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment answered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/comments/{comment}/specialize": {
            "post": {
                "tags": [
                    "Admin - Comments"
                ],
                "summary": "Mark comment as special",
                "operationId": "adminCommentsSpecialize",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment specialized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StateResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tehran"
                    },
                    "slug": {
                        "type": "string",
                        "example": "tehran"
                    },
                    "code": {
                        "type": "string",
                        "example": "THR",
                        "nullable": true
                    },
                    "cities_count": {
                        "type": "integer",
                        "example": 16,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CityResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "state_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Karaj"
                    },
                    "slug": {
                        "type": "string",
                        "example": "alborz-karaj"
                    },
                    "code": {
                        "type": "string",
                        "example": "KRJ",
                        "nullable": true
                    },
                    "is_capital": {
                        "type": "boolean",
                        "example": true
                    },
                    "state": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/StateResource"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Use Bearer {token} obtained from /auth/verify."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}